Update Linux release action
