+++ .github/workflows/linux_release.yml
@@ -11,19 +11,16 @@ env:
   Python_REQUIRED_VERSION: 3.8
 
 jobs:
-  linux_release:
+  build_installer_artifact:
+    name: Build Installer and Upload  # keeping upload in this job so we could potentially download problematic builds
     runs-on: ${{ matrix.os }}
-    continue-on-error: ${{ matrix.allow_failure }}
+    continue-on-error: false
     strategy:
       # fail-fast: Default is true, switch to false to allow one platform to fail and still run others
       fail-fast: false
       matrix:
         os: [ubuntu-20.04, ubuntu-22.04]
-        include:
-        - os: ubuntu-20.04
-          allow_failure: false
-        - os: ubuntu-22.04
-          allow_failure: false
+
     permissions:
       # Needed permission to upload the release asset
       contents: write
@@ -58,6 +55,8 @@ jobs:
     - name: Create Build Directory
       run: cmake -E make_directory ./build/
 
+    # should we pin to a specific version of CMake here like we did on Windows?
+
     - name: Configure CMake
       working-directory: ./build
       shell: bash
@@ -73,6 +72,15 @@ jobs:
       working-directory: ./build
       shell: bash
       run: cmake --build . --target package -j 4
+
+    - name: Upload Tarball as artifact for testing
+      uses: actions/upload-artifact@v4
+      with:
+        name: energyplus-${{ matrix.os }}
+        path: build/EnergyPlus-*-x86_64.tar.gz
+        if-no-files-found: error
+        retention-days: 7
+        overwrite: false
 
     - name: Upload Tarball to release
       uses: svenstaro/upload-release-action@v2
@@ -101,3 +109,54 @@ jobs:
         overwrite: true
         file_glob: true
         asset_name: ${{ matrix.os }}_LinuxShellInstaller
+
+
+  test_package:
+    name: Test Built Package
+    needs: build_installer_artifact
+    runs-on: ${{ matrix.os }}
+    strategy:
+      # fail-fast: Default is true, switch to false to allow one platform to fail and still run others
+      fail-fast: false
+      matrix:
+        name: [20, 22]
+        include:
+          - name: 20
+            os: ubuntu-20.04
+            test_key: ubuntu2004
+          - name: 22
+            os: ubuntu-22.04
+            test_key: ubuntu2204
+
+    steps:
+      - uses: actions/checkout@v3  # Still need E+ checked out to get testing scripts
+        with:
+          path: checkout
+
+      - name: Set up Python ${{ env.Python_REQUIRED_VERSION }}
+        uses: actions/setup-python@v4
+        id: setup-python
+        with:
+          python-version: ${{ env.Python_REQUIRED_VERSION }}
+          architecture: ${{ matrix.arch }}
+
+      - name: Gather Test Package from Artifacts
+        uses: actions/download-artifact@v4
+        with:
+          name: energyplus-${{ matrix.os }}
+          path: package
+
+      - name: Check Contents
+        shell: bash
+        run: ls
+
+      - name: Check Package contents
+        shell: bash
+        working-directory: package
+        run: ls
+
+      #      - name: Setup tmate session
+      #        uses: mxschmitt/action-tmate@v3
+
+      - name: Run Package Tests
+        run: python checkout/scripts/package_tests/runner.py --verbose ${{ matrix.test_key }} package/
